Backend/requirements.txt
Django>=5.0.0
django-cors-headers
drf-yasg
psycopg2-binary
django-redis
python-dotenv
dj-database-url
whitenoise
requests
rest_framework_simplejwt
djangorestframework==3.14.0
setuptools>=65.0.0
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
django-allauth==0.57.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
PyMySQL==1.1.0
python-decouple==3.8
django-extensions==3.2.3
celery==5.3.4
redis==5.0.1
Pillow==10.1.0
setuptools>=65.0.0
drf-spectacular
drf-spectacular-sidecar
